[perl #34935] [PATCH] r7818: removing more warnings in src/
this patch against r7818 should eliminate many more warning messages in
src/*.c

[perl #34936] [PATCH] r7818: removing more warnings in other dirs

this patch covers charset/, ops/, pf/, classes/, and io/
adds typecast info, and removes unreferenced locals

Courtesy of Jerry Gay <jerry.gay@gmail.com>


git-svn-id: https://svn.parrot.org/parrot/trunk@7822 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ io/io_stdio.c
@@ -415,7 +415,7 @@ PIO_stdio_seek(theINTERP, ParrotIOLayer *layer, ParrotIO *io,
 
     errno = 0;
 
-    if ((pos = fseek((FILE*)io->fd, offset, whence)) >= 0) {
+    if ((pos = fseek((FILE*)io->fd, (long)offset, whence)) >= 0) {
         io->lpos = io->fpos;
         io->fpos = pos;
     }
